# docs/specs/S01-openapi.yaml
openapi: 3.1.0
info:
  title: Baby Growth Record API
  version: 2.3.0
  description: API specification for the baby growth record SaaS platform. IDs are decimal strings at API boundary and stored as bigint in database; time fields use UTC millisecond timestamps.
servers:
  - url: /api/v1
...
paths:
...
  /events:
    get:
      tags: [events]
      summary: List growth events
      parameters:
        - in: query
          name: family_id
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - in: query
          name: baby_id
          schema:
            type: string
            pattern: '^[0-9]+$'
        - in: query
          name: event_type
          schema:
            $ref: '#/components/schemas/EventType'
        - in: query
          name: date_from
          schema:
            type: integer
            format: int64
        - in: query
          name: date_to
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
        '400':
          description: invalid id or query argument
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidArgumentResponse'
              example:
                code: INVALID_ARGUMENT
                message: invalid id format
                data:
                  field: family_id
                  reason: must match ^[0-9]+$
                  value: abc
    post:
      tags: [events]
      summary: Create growth event
      parameters:
        - in: header
          name: X-Idempotency-Key
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreateRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
  /events/{event_id}:
    get:
      tags: [events]
      summary: Get event detail
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
    patch:
      tags: [events]
      summary: Update event
      parameters:
        - $ref: '#/components/parameters/EventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventUpdateRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
    delete:
      tags: [events]
      summary: Soft delete event
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
...
  /reports/export:
    post:
      tags: [reports]
      summary: Create export task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportTaskCreateRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportTaskResponse'
  /reports/exports:
    get:
      tags: [reports]
      summary: List export tasks
      parameters:
        - in: query
          name: family_id
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
  /reports/exports/{task_id}:
    get:
      tags: [reports]
      summary: Get export task detail
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportTaskResponse'
...
components:
  parameters:
...
    EventId:
      in: path
      name: event_id
      required: true
      schema:
        type: string
        pattern: '^[0-9]+$'
  schemas:
    CommonResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: object
          additionalProperties: true
    InvalidArgumentData:
      type: object
      properties:
        field:
          type: string
        reason:
          type: string
        value:
          type: string
    InvalidArgumentResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/InvalidArgumentData'
...
    EventType:
      type: string
      enum: [feeding, excretion, measurement, sleep, medication, vaccine, milestone]
...
    EventPayload:
      type: object
      additionalProperties: true
      description: Type-specific payload for growth events.
    EventCreateRequest:
      type: object
      required: [family_id, baby_id, event_type, occurred_at, payload]
      properties:
        family_id:
          type: string
          pattern: '^[0-9]+$'
        baby_id:
          type: string
          pattern: '^[0-9]+$'
        event_type:
          $ref: '#/components/schemas/EventType'
        occurred_at:
          type: integer
          format: int64
        start_at:
          type: integer
          format: int64
        end_at:
          type: integer
          format: int64
        timezone:
          type: string
          default: Asia/Shanghai
        notes:
          type: string
        payload:
          $ref: '#/components/schemas/EventPayload'
    EventUpdateRequest:
      type: object
      properties:
        occurred_at:
          type: integer
          format: int64
        start_at:
          type: integer
          format: int64
        end_at:
          type: integer
          format: int64
        notes:
          type: string
        payload:
          $ref: '#/components/schemas/EventPayload'
    Event:
      type: object
      properties:
        id:
          type: string
          pattern: '^[0-9]+$'
        family_id:
          type: string
          pattern: '^[0-9]+$'
        baby_id:
          type: string
          pattern: '^[0-9]+$'
        event_type:
          $ref: '#/components/schemas/EventType'
        occurred_at:
          type: integer
          format: int64
        start_at:
          type: integer
          format: int64
        end_at:
          type: integer
          format: int64
        timezone:
          type: string
        notes:
          type: string
        payload:
          $ref: '#/components/schemas/EventPayload'
        status:
          type: string
          enum: [active, deleted]
    EventResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Event'
    EventListResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Event'
...
    ExportTaskCreateRequest:
      type: object
      required: [family_id, baby_id, report_type, date_from, date_to]
      properties:
        family_id:
          type: string
          pattern: '^[0-9]+$'
        baby_id:
          type: string
          pattern: '^[0-9]+$'
        report_type:
          type: string
          enum: [daily, weekly, monthly, custom]
        date_from:
          type: integer
          format: int64
        date_to:
          type: integer
          format: int64
    ExportTaskResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                id:
                  type: string
                  pattern: '^[0-9]+$'
                status:
                  type: string
                  enum: [pending, running, succeeded, failed]
                download_url:
                  type: string
                expires_at:
                  type: integer
                  format: int64
...